Develop

+++ api/controllers/user.controller.js
@@ -47,8 +47,8 @@ async function getUserMedCert(req, res) {
   try {
     const user = await UserModel.findById(req.params.id)
     const cert = user.studentData.medCert
-    if(!cert) return res.status(400).send('No Medical Certificate')
-    res.status(200).sendFile(cert, {root: 'public/med-certs'})
+    if (!cert) return res.status(400).send('No Medical Certificate')
+    res.status(200).sendFile(cert, { root: 'public/med-certs' })
   } catch (error) {
     res.status(500).send(`Request Error: ${error}`)
   }
@@ -68,15 +68,6 @@ async function getUserDriveLic(req, res) {
 async function updateUser(req, res) {
   try {
     const user = await UserModel.findByIdAndUpdate(req.params.id, req.body, { password: 0 })
-    if (user.role === 'student' && req.body.hasOwnProperty('studentData.teacher')) {
-      const teacher = await UserModel.findById(Object.values(req.body)[0])
-      if (teacher.teacherData.students.indexOf(user.id) !== -1) {
-        return res.send('student already assigned')
-      } else {
-        teacher.teacherData.students.push(req.params.id)
-        teacher.save()
-      }
-    }
     res.status(200).json({ message: `${user.name}'s profile updated!`, user })
   } catch (error) {
     res.status(500).send(`Request Error: ${error}`)
@@ -85,8 +76,30 @@ async function updateUser(req, res) {
 
 async function deleteUser(req, res) {
   try {
-    const user = await UserModel.findByIdAndDelete(req.params.id)
-    res.status(200).send(`${user.name}'s profile deleted`)
+    const user = await UserModel.findById(req.params.id)
+    const admin = res.locals.user.id
+    if (user.id === admin) return res.send('Dear, only God could delete yourself from this world')
+
+    if (user.studentData.teacher || user.teacherData.student) {
+      return res.send(`User ${user.name} has teacher/student assigned and cannot be deleted`)
+    } else {
+      const delUser = await UserModel.findByIdAndDelete(req.params.id)
+      res.status(200).send(`${delUser.name}'s profile deleted`)
+    }
+  } catch (error) {
+    res.status(500).send(`Request Error: ${error}`)
+  }
+}
+
+async function assignTeacher(req, res) {
+  try {
+    const student = await UserModel.findById(req.params.studentId)
+    const teacher = await UserModel.findById(req.params.teacherId)
+    student.studentData.teacher = req.params.teacherId
+    teacher.teacherData.students.push(student.id)
+    await student.save()
+    await teacher.save()
+    res.status(200).send(`Teacher ${teacher.name} assigned to ${student.name}`)
   } catch (error) {
     res.status(500).send(`Request Error: ${error}`)
   }
@@ -113,8 +126,8 @@ async function updateMyProfile(req, res) {
 function getProfilePhoto(req, res) {
   try {
     const photo = res.locals.user.photo
-    if(!photo) return res.status(400).send('No profile photo')
-    res.status(200).sendFile(photo, {root: 'public/profile-photos'})
+    if (!photo) return res.status(400).send('No profile photo')
+    res.status(200).sendFile(photo, { root: 'public/profile-photos' })
   } catch (error) {
     res.status(500).send(`Request error: ${error}`)
   }
@@ -123,8 +136,8 @@ function getProfilePhoto(req, res) {
 function getMyMedCert(req, res) {
   try {
     const cert = res.locals.user.studentData.medCert
-    if(!cert) return res.status(400).send('No Medical Certificate')
-    res.status(200).sendFile(cert, {root: 'public/med-certs'})
+    if (!cert) return res.status(400).send('No Medical Certificate')
+    res.status(200).sendFile(cert, { root: 'public/med-certs' })
   } catch (error) {
     res.status(500).send(`Request error: ${error}`)
   }
@@ -173,20 +186,20 @@ async function createPractice(req, res) {
       teacherUser.teacherData.lessons.push(practice.id)
       teacherUser.save()
 
-      res.status(200).json({ message: `You've booked a drivning lesson!`, practice })
+      res.status(200).json({ message: `You've booked a driving lesson!`, practice })
     } else {
       const dayPractices = await DriveLessonModel.find({ date: req.body.date, teacher: teacher })
       const booked = dayPractices.map(practice => {
         return practice.bookSlot
       })
-      const arr = ["09:00", "10:00","11:00","12:00","13:00","16:00","17:00","18:00"]
+      const arr = ["9:00", "10:00","11:00","12:00","13:00","16:00","17:00","18:00"]
       const free = arr.filter(hour => {
         return booked.indexOf(hour) === -1
       })
       if (free.length > 0) {
         res.status(200).send(`The teacher is busy!!! Available hours for this date are: ${free}`)
       } else {
-        res.status(200).send('Sorry, currently there are no more slots available por this date')
+        res.status(200).send('Sorry, currently there are no more slots available for this date')
       }
     }
   } catch (error) {
@@ -273,6 +286,7 @@ module.exports = {
   getUserDriveLic,
   updateUser,
   deleteUser,
+  assignTeacher,
   getMyProfile,
   updateMyProfile,
   getProfilePhoto,
